+++ hutool-core/src/test/java/cn/hutool/core/date/DateUtilTest.java
@@ -1,7 +1,7 @@
 package cn.hutool.core.date;
 
 import cn.hutool.core.collection.CollUtil;
-import cn.hutool.core.date.BetweenFormater.Level;
+import cn.hutool.core.date.BetweenFormatter.Level;
 import cn.hutool.core.date.format.FastDateFormat;
 import cn.hutool.core.util.RandomUtil;
 import org.junit.Assert;
